+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ziti CLI
     * The CLI has been cleaned up and unused, unusable and underused components have been removed or hidden
     * Add create/delete transit-router CLI commands
+    * [Issue-706](https://github.com/openziti/ziti/issues/706) - Add port check to quickstart
 
 ## Ziti CLI